added utils file with fetch models functions and adjusted type
