+++ biosimulator_processes/processes/copasi_process.py
@@ -16,6 +16,7 @@ from basico import (
     biomodels
 )
 from process_bigraph import Process, Composite, pf
+from biosimulator_processes.utils import fetch_biomodel
 
 
 # 1. Map config_schema params to SEDML syntax/semantics (ie: 'model_file')
@@ -24,28 +25,6 @@ from process_bigraph import Process, Composite, pf
     # and provides num iterations and parameter in model_changes
 
 # define config schema type decs here
-
-
-# TODO put in utils / SED-related
-def _fetch_biomodel(term: str, index: int = 0):
-    """Search for models matching the term and return an instantiated model from BioModels.
-
-        Args:
-            term:`str`: search term
-            index:`int`: selector index for model choice
-
-        Returns:
-            `CDataModel` instance of loaded model.
-        TODO: Implement a dynamic search of this
-    """
-    models = biomodels.search_for_model(term)
-    model = models[index]
-    sbml = biomodels.get_content_for_model(model['id'])
-    return load_model_from_string(sbml)
-
-
-def fetch_biomodel(model_id: str):
-    return biomodels.get_content_for_model(model_id)
 
 
 MODEL_CHANGES_TYPE = {
